--- a/Cargo.toml
+++ b/Cargo.toml
@@ -31,3 +31,4 @@ tower = { version = "0.4.13", features = ["default"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 base64 = "0.22.1"
+regex = "1.10.5"