--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod components;
+pub mod middlewares;
+pub mod routes;
+pub mod scraper;