Here's the scraper! I've added it in a few parts:

- `src/bin/scraper.rs` is the CLI interface which scrapes Banner and spits out SQLite databases
- `src/routes.rs` has the new `DatabaseAppState`, which opens a new read-only SQLite connection for each request.
- `src/scraper.rs` has the types and functions that can be reused between the two parts.

As you may have noticed, I also restructured the project into a single `src/lib.rs`, and the multiple binary targets in `src/bin/`.

Since this works now, I don't recommend that we keep using the `AppState` trait, and instead directly use the `DatabaseAppState` type. Unless this is for some long-term dependency-injection plan, I think it's better to just provide a SQLite connection to each handler and let queries be done ad-hoc.

Closes #7