@brennanmcmicking 2024-07-30
Just one over-arching question: is there a reason that we can't open 1 SQLite connection and pass it around to the routes rather than opening a new connection per-route? Seems like this would add a bit of overhead even in the "db-is-stored-on-disk" case.