@malcolmseyd 2024-07-30
> Just one over-arching question: is there a reason that we can't open 1 SQLite connection and pass it around to the routes rather than opening a new connection per-route? Seems like this would add a bit of overhead even in the "db-is-stored-on-disk" case.

`rusqlite::Connection` is `!Sync` so you can't share between threads at all (i.e. `Arc<Mutex<Connection>>` doesn't satisfy trait bounds for `S` in `Router<S>`). [This great StackOverflow post](https://stackoverflow.com/questions/52801244/compiler-says-that-data-cannot-be-shared-between-threads-safely-even-though-the) explains it's because they use `RefCell` internally. The most correct solution here would be to use [`r2d2`](https://lib.rs/crates/r2d2) to pool connections, I could do that if you'd like.