@malcolmseyd 2024-07-30
Okay I decided to add the connection pooler and load tested the release build before/after. It made a nice improvement, especially in the tail latencies!

```console
$ # before pooler
$ echo "GET http://127.0.0.1:8080" | vegeta attack -duration=10s -rate=5000 -max-workers=1000 | vegeta report
Requests      [total, rate, throughput]         50000, 5000.06, 4999.78
Duration      [total, attack, wait]             10s, 10s, 557.391µs
Latencies     [min, mean, 50, 90, 95, 99, max]  167.845µs, 567.077µs, 406.848µs, 880.309µs, 1.175ms, 3.479ms, 21.206ms
Bytes In      [total, mean]                     1124100000, 22482.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:50000

$ # after pooler
$ echo "GET http://127.0.0.1:8080" | vegeta attack -duration=10s -rate=5000 -max-workers=1000 | vegeta report
Requests      [total, rate, throughput]         50000, 5000.07, 4999.98
Duration      [total, attack, wait]             10s, 10s, 166.615µs
Latencies     [min, mean, 50, 90, 95, 99, max]  103.433µs, 367.407µs, 307.136µs, 661.601µs, 721.852µs, 843.554µs, 2.909ms
Bytes In      [total, mean]                     1124100000, 22482.00
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:50000
```